+++ core/src/main/groovy/com/ainrif/apiator/core/Apiator.groovy
@@ -15,10 +15,7 @@
  */
 package com.ainrif.apiator.core
 
-import com.ainrif.apiator.core.model.api.ApiContext
-import com.ainrif.apiator.core.model.api.ApiEndpoint
-import com.ainrif.apiator.core.model.api.ApiScheme
-import com.ainrif.apiator.core.model.api.ClientApiInfo
+import com.ainrif.apiator.core.model.api.*
 import com.ainrif.apiator.doclet.ApiatorDoclet
 import com.ainrif.apiator.doclet.SourcePathDetector
 import com.ainrif.apiator.doclet.javadoc.DocletInfoIndexer
@@ -69,9 +66,7 @@ class Apiator {
             Set<Class<?>> apiClasses = scanForApi()
 
             apiClasses
-                    .findResults {
-                        return config.provider.getContextStack(it)
-                    }
+                    .findResults { config.provider.getContextStack(new ApiType(it)) }
                     .collect { ctxStack ->
                         def apiCtx = new ApiContext(
                                 name: ctxStack.name,
@@ -142,7 +137,7 @@ class Apiator {
                     .findAll { !new File(it).exists() }
                     .each { log.warn('There are no source path like {}', it) }
 
-            def docletClassPath = null
+            String docletClassPath = null
             if (extraClassPath) {
                 docletClassPath = extraClassPath.findAll { new File(it.file).exists() }
                         .collect { it.file }
@@ -154,12 +149,13 @@ class Apiator {
                 return new DocletInfoIndexer([:])
             }
 
-            def result = ApiatorDoclet.runDoclet(dConf.sourcePath, docletClassPath, dConf.includeBasePackage, null)
+            def result = ApiatorDoclet.runDoclet(dConf.sourcePath, docletClassPath, dConf.includeBasePackage)
             if (!result.success) {
                 log.warn('Execution of doclet was unsuccessful. Doc source will be skipped')
                 return new DocletInfoIndexer([:])
             }
 
+            // TODO katoquro: 1/10/19 consider using of jackson parser to replace compile dynamic for mapping
             def filePath = result.outputFile
             def javaDocInfo = new JsonSlurper().parse(new File(filePath)) as JavaDocInfo