+++ core/src/main/groovy/com/ainrif/apiator/doclet/javadoc/MethodMergedInfo.groovy
@@ -19,9 +19,11 @@ package com.ainrif.apiator.doclet.javadoc
 import com.ainrif.apiator.core.model.api.ApiEndpointParam
 import com.ainrif.apiator.doclet.model.MethodInfo
 import com.ainrif.apiator.doclet.model.ParamInfo
+import groovy.transform.CompileDynamic
 
 import javax.annotation.Nullable
 
+@CompileDynamic
 class MethodMergedInfo {
     private List<MethodInfo> methodInfos