+++ provider/micronaut-provider/src/main/groovy/com/ainrif/apiator/provider/micronaut/MicronautMethodStack.groovy
@@ -21,6 +21,7 @@ import com.ainrif.apiator.core.model.api.ApiEndpointParamType
 import com.ainrif.apiator.core.model.api.ApiType
 import com.ainrif.apiator.core.reflection.MethodStack
 import com.ainrif.apiator.core.reflection.RUtils
+import groovy.transform.CompileDynamic
 import io.micronaut.http.HttpHeaders
 import io.micronaut.http.HttpParameters
 import io.micronaut.http.HttpRequest
@@ -63,13 +64,17 @@ class MicronautMethodStack extends MethodStack {
         return this.last().name
     }
 
+    @CompileDynamic
     @Override
     String getPath() {
         Annotation pathOpt = ENDPOINT_PATH_ANNOTATIONS.findResult {
             AnnotationUtils.findAnnotation(this.last(), it)
         }
 
-        return pathOpt?.value() ?: pathOpt?.uri() ?: '/'
+        def value = pathOpt.value() as String
+        def uri = pathOpt.uri() as String
+
+        return (value.length() <=> uri.length()) >= 0 ? value : uri
     }
 
     @Override
@@ -84,6 +89,7 @@ class MicronautMethodStack extends MethodStack {
     /**
      * https://docs.micronaut.io/latest/guide/index.html#_variables_resolution
      */
+    @CompileDynamic
     @Override
     List<ApiEndpointParam> getParams() {
         String rawPath = this.path
@@ -219,7 +225,7 @@ class MicronautMethodStack extends MethodStack {
         result.addAll(remainingParamTypes)
 
         if (expandedBodyParams) {
-            def typeName = 'adhoc_type.' + context.last().simpleName + getName().capitalize()
+            def typeName = 'adhoc_type.' + context.apiType.rawType.simpleName + getName().capitalize()
 
             def genClass = new ByteBuddy()
                     .makeInterface()
@@ -242,7 +248,10 @@ class MicronautMethodStack extends MethodStack {
                     index: -1,
                     name: null,
                     type: new ApiType(generated),
-                    httpParamType: ApiEndpointParamType.BODY))
+                    httpParamType: ApiEndpointParamType.BODY,
+                    // TODO katoquro: 1/10/19 think about combining annotations from params
+                    annotations: []
+            ))
         }
 
         return result