+++ renderer/core-json-renderer/plugin/jax-rs-plugin-core-json-renderer/src/main/groovy/com/ainrif/apiator/renderer/plugin/jaxrs/JaxRsCompositePlugin.groovy
@@ -21,7 +21,7 @@ import com.ainrif.apiator.renderer.plugin.spi.CoreJsonRendererPlugin
 
 class JaxRsCompositePlugin implements CompositePlugin {
     @Override
-    List<CoreJsonRendererPlugin> getPlugins() {
-        return [new JaxRsModelTypePlugin(), new JaxRsParamPlugin()]
+    List<? extends CoreJsonRendererPlugin> getPlugins() {
+        return [new JaxRsModelTypePlugin(), new JaxRsParamPlugin()] as List<CoreJsonRendererPlugin>
     }
 }