+++ renderer/core-json-renderer/plugin/spi-plugin-core-json-renderer/src/main/java/com/ainrif/apiator/renderer/plugin/spi/CompositePlugin.java
@@ -22,5 +22,5 @@ import java.util.List;
  * interface to pass a bundle of plugins as one plugin
  */
 public interface CompositePlugin extends CoreJsonRendererPlugin {
-    List<CoreJsonRendererPlugin> getPlugins();
+    List<? extends CoreJsonRendererPlugin> getPlugins();
 }